0.8 client decode jwt (#314)
* Only set app token if it’s not expired.

Modifies the getJWT function to return undefined if the token has expired.  This gives you a quick way of knowing if the user is (more than likely) authenticated without having to make a request.

* Allow client to decode token.

This adds an `app.decodeJWT` function that decodes a token synchronously.  With a non-expired token, it returns the token payload, otherwise it returns undefined.

Will need docs before publishing 0.8 or whichever version we’re doing next.

* adding instanbul code coverage

* Remove permissions hooks and middleware which will be put into feathers-permissions (#307)

* Remove ‘user’ since server no longer supports it.

* Make API better match the server.

The only difference is that verifyJWT and getJWT are sync, so they don’t return promises.