client: use Authentication class, make `getJWT` and `verifyJWT` async (#317)
* client: make `getJWT` and `verifyJWT` async

This modifies both `getJWT` and `verifyJWT` to return promises.  `getJWT` has also been updated to check the store for the token again to support environments that don’t load cookies immediately.

* Removed unused var.

* Prevent saving ‘null’ or ‘undefined’ into storage

Some forms of webStorage plugins, when you pass `null` or `undefined` will actually stringify it into `’null’` or `’undefined’`, which results in a malformed jwt error when `verifyJWT` is run.  This prevents that.

* Started authentication class

* refactor to use Authentication class.