Allow the ability to specify a custom service to be used
When passing the config to the configure function for authentication you can provide a function, similar to how most services are created, that registers a custom service. This will not register the original, default, service that is normal created; it will instead set the `authentication` service to the one specified under the `service` property within the config.